+++ esibd/devices/ni9263/ni9263.py
@@ -90,4 +90,5 @@ class VoltageController(DeviceController):
 
     def closeCommunication(self) -> None:
         super().closeCommunication()
+        # nothing to close
         self.initialized = False